llm-proxy/pyproject.toml
[project]
name = "llm-proxy"
version = "1.0.0"
description = "Multi-vendor LLM API Proxy Server"
requires-python = ">=3.12"
dependencies = [
    # Web Framework
    "fastapi>=0.116.1",
    "fastapi-jwt-auth>=0.5.0",
    "uvicorn>=0.23.1",
    "gunicorn>=23.0.0",
    "httptools>=0.6.4",
    "pydantic>=2.11.7",
    "python-multipart>=0.0.9",

    # HTTP & Serialization
    "httpx>=0.28.1",
    "requests>=2.31.0",
    "orjson>=3.11.1",

    # Environment
    "python-dotenv>=1.1.0",

    # AI Provider SDK
    "openai>=1.97.1",
    "anthropic>=0.60.0",
    "google-genai>=1.55.0",
    "aioboto3>=12.3.0",
    "xai-sdk>=1.4.0",

    # LLM Utilities
    "tiktoken>=0.9.0",
    "numpy>=2.2.6",
    "fastmcp>=2.10.6",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
